# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.CreatePostRequest:
    properties:
      content:
        maxLength: 1000
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 200
        type: string
    required:
    - content
    - title
    type: object
  request.UpdatePostRequest:
    properties:
      content:
        maxLength: 50000
        minLength: 1
        type: string
      tags:
        items:
          type: string
        maxItems: 20
        type: array
      title:
        maxLength: 500
        minLength: 1
        type: string
    type: object
  response.CommentResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      user:
        $ref: '#/definitions/response.UserResponse'
      user_id:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.PostResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/response.CommentResponse'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
      version:
        type: integer
    type: object
  response.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Social network REST API
  title: Social API
  version: "1.0"
paths:
  /posts/:
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Create post request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a post
      tags:
      - Posts
  /posts/{postID}/:
    delete:
      description: Delete a post by ID
      parameters:
      - description: Post ID
        format: int64
        in: path
        name: postID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a post
      tags:
      - Posts
    get:
      description: Get a post by ID
      parameters:
      - description: Post ID
        format: int64
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PostResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a post
      tags:
      - Posts
    patch:
      consumes:
      - application/json
      description: Update a post by ID
      parameters:
      - description: Post ID
        format: int64
        in: path
        name: postID
        required: true
        type: integer
      - description: Update post request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update a post
      tags:
      - Posts
  /posts/feed:
    get:
      description: Get paginated posts for the current user
      parameters:
      - description: Number of posts to return
        in: query
        name: limit
        type: integer
      - description: Number of posts to skip
        in: query
        name: offset
        type: integer
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Search keyword
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: Filter by tags
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.PostResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get feed
      tags:
      - Posts
  /users/{userID}:
    get:
      description: Get user information by user ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get user by ID
      tags:
      - Users
  /users/{userID}/follow:
    delete:
      description: Stop following another user by user ID
      parameters:
      - description: User ID to unfollow
        format: int64
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Successfully unfollowed user
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Unfollow a user
      tags:
      - Followers
    post:
      description: Follow another user by user ID
      parameters:
      - description: User ID to follow
        format: int64
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Successfully followed user
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Follow a user
      tags:
      - Followers
swagger: "2.0"